calculate bearing/range from request origin to forecast point

+++ data-collection-service/src/github.com/buskersguidetotheuniverse.org/types/conditions.go
@@ -60,8 +60,8 @@ type Geometry struct {
 }
 
 type Distance struct {
-	Bearing     int
-	Range       int
+	Bearing     float64
+	Range       float64
 	RangeUnit   string
 	BearingUnit string
 }
